Increase the minimum word length to 3 letters
The prior code was inconsistently documented with some parts specifing a
3 letter limit and others a 2 letter one.  For example, the docs for
pplen incorrectly specified that the minimum value was 3 when
RandPassGenerator.java allowed 2.

The source of this confusion is likely due to the fact that WordSet.java
used a default length of 3.  This was, however, being overridden in all
circumstances by the setLengthRange call immediately following
construction in RandPassGenerator.java.

This change unifies everything on three letters under the belief that
the vast majority of the existing two letter words are hard to remember
and relatively unknown.